requirements.txt
uvicorn==0.35.0
fastapi==0.116.1
starlette==0.47.2
pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.1
annotated-types==0.7.0
motor==3.7.1
pymongo==4.14.0
dnspython==2.7.0
python-jose==3.5.0
bcrypt==4.3.0
email_validator==2.2.0
jinja2==3.1.4
aiosmtplib==3.0.1
aio-pika==9.4.1
python-dotenv==1.1.1
art==6.1